Update README.md
## Explanation

You may have already seen the popular visualization of the fourth spatial dimension on computer/in VR known as 4D Toys. To get an excellent demonstration and explanation of what the fourth spatial dimension actually means, you can check out [this video](https://www.youtube.com/watch?v=0t4aKJuKP0Q) posted by the creator.

The basis of 4D Room is different. 4D Toys allows you to see the _intersection_ of a 4D object with 3D space. 4D Room aims to show the _projection_ of a 4D object into 3D space. To make the difference clearer, here is a picture showing the intersection and the projection of a 3D cube in a 2D plane.

As you can see, the projection gives you a much clearer understanding of what the two-dimensional drawing represents. This is because the projection gives a sense of depth in an additional dimension.

In this project, we begin with representing the shapes mathematically in 4D, simulate the rotations of these shapes, and then display them in the appropriate number of dimensions -- two dimensions if you are on a screen, and three dimensions if you are wearing a VR headset.